Created new components for ArticleFeedInfo and ArticleStatInfo, Fixes for Teacher Preview
