+++ src/components/ArticleStatInfo.js
@@ -1,0 +1,24 @@
+import * as s from "./ArticleStatInfo.sc";
+import { getStaticPath } from "../utils/misc/staticPath";
+import { estimateReadingTime } from "../utils/misc/readableTime";
+
+export default function ArticleStatInfo({ cefr_level, articleInfo }) {
+  return (
+    <s.StatContainer>
+      <s.Difficulty>
+        <img
+          src={getStaticPath("icons", cefr_level + "-level-icon.png")}
+          alt="difficulty icon"
+        ></img>
+        <span>{cefr_level}</span>
+      </s.Difficulty>
+      <s.ReadingTimeContainer>
+        <img
+          src={getStaticPath("icons", "read-time-icon.png")}
+          alt="read time icon"
+        ></img>
+        <span>~ {estimateReadingTime(articleInfo.metrics.word_count)}</span>
+      </s.ReadingTimeContainer>
+    </s.StatContainer>
+  );
+}
